requirements.txt
requests
matplotlib
Pillow
colorama
pytesseract
baidu-aip